src/bin/shipyard_airflow/requirements-direct.txt
alembic
apache-airflow[crypto,celery,hive,hdfs,jdbc,postgres]==3.0.2
apache-airflow-providers-standard
arrow
celery
Flask
Flask-AppBuilder
graphviz
pytz
falcon
jsonschema
networkx
kombu
kubernetes
marshmallow-sqlalchemy
mock
psycopg2-binary
pyarrow
pylibyaml
python3-memcached
PyYAML
regex
requests
responses
setuptools
SQLAlchemy
tabulate
ulid
uWSGI
Werkzeug
WTForms
itsdangerous
pandas

# Openstack Caracal 2024.1
# https://releases.openstack.org/caracal/index.html
barbican==18.0.0

python-barbicanclient==5.7.0
python-keystoneclient==5.4.0

keystoneauth1==5.6.0
keystonemiddleware==10.6.0

oslo.cache==3.7.0
oslo.concurrency==6.0.0
oslo.config==9.4.0
oslo.context==5.5.0
oslo.db==15.0.0
oslo.i18n==6.3.0
oslo.log==5.5.1
oslo.messaging==14.7.2
oslo.metrics==0.8.0
oslo.middleware==6.1.0
oslo.policy==4.3.0
oslo.serialization==5.4.1
oslo.service==3.4.1
oslo.upgradecheck==2.3.0
oslo.utils==7.1.0
oslo.versionedobjects==3.3.0


# Airship dependencies
git+https://opendev.org/airship/deckhand.git@d8cc9f3ed2c9f2b9d9bf271a72c9bf301f16f647#egg=deckhand
git+https://opendev.org/airship/drydock.git@1886a2f09fbd6a06fa4c5da86361b255edd46b11#egg=drydock_provisioner&subdirectory=python
git+https://opendev.org/airship/armada.git@61d1659d4276a4be5e4711d803c14b721d910b85#egg=armada
git+https://opendev.org/airship/promenade.git@63bf9dd5693b6e1bf372addadab703cade80cce3#egg=promenade
